moves compression tests from integration ones to acceptance
